Don't add the link if it already exists
When I use browser history to to back/forward on github, every time I see the main repo page, it creates a new link below the others. This fixes that--it switches to using an id instead of a class, and it'll only make an element if there isn't one already.

I'm using chrome with tampermonkey if that makes a difference.